order/requirements.txt
Flask[async]==3.0.2
redis==5.0.3
gunicorn==21.2.0
msgspec==0.18.6
requests==2.31.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
flask-sqlalchemy==3.1.1
kafka-python==2.0.5
debugpy==1.8.12
uvicorn==0.34.0
asgiref==3.8.1